Convert floats to doubles

+++ embedded/include/limesuiteng/embedded/lms7002m/lms7002m.h
@@ -48,8 +48,8 @@ typedef struct lms7002m_hooks {
 struct lms7002m_context* lms7002m_create(const lms7002m_hooks* hooks);
 void lms7002m_destroy(struct lms7002m_context* context);
 
-float lms7002m_get_reference_clock(struct lms7002m_context* context);
-lime_Result lms7002m_set_reference_clock(struct lms7002m_context* context, float frequency_Hz);
+double lms7002m_get_reference_clock(struct lms7002m_context* context);
+lime_Result lms7002m_set_reference_clock(struct lms7002m_context* context, double frequency_Hz);
 
 lime_Result lms7002m_enable_channel(struct lms7002m_context* self, const bool isTx, const uint8_t channel, const bool enable);
 
@@ -61,26 +61,26 @@ lime_Result lms7002m_reset_logic_registers(struct lms7002m_context* self);
 
 lime_Result lms7002m_tune_cgen_vco(struct lms7002m_context* self);
 
-lime_Result lms7002m_set_frequency_cgen(struct lms7002m_context* context, float frequency_Hz);
-float lms7002m_get_frequency_cgen(struct lms7002m_context* self);
+lime_Result lms7002m_set_frequency_cgen(struct lms7002m_context* context, double frequency_Hz);
+double lms7002m_get_frequency_cgen(struct lms7002m_context* self);
 
-lime_Result lms7002m_set_rbbpga_db(struct lms7002m_context* self, const float value, const uint8_t channel);
-float lms7002m_get_rbbpga_db(struct lms7002m_context* self, const uint8_t channel);
+lime_Result lms7002m_set_rbbpga_db(struct lms7002m_context* self, const double value, const uint8_t channel);
+double lms7002m_get_rbbpga_db(struct lms7002m_context* self, const uint8_t channel);
 
-lime_Result lms7002m_set_rfelna_db(struct lms7002m_context* self, const float value, const uint8_t channel);
-float lms7002m_get_rfelna_db(struct lms7002m_context* self, const uint8_t channel);
+lime_Result lms7002m_set_rfelna_db(struct lms7002m_context* self, const double value, const uint8_t channel);
+double lms7002m_get_rfelna_db(struct lms7002m_context* self, const uint8_t channel);
 
-lime_Result lms7002m_set_rfe_loopback_lna_db(struct lms7002m_context* self, const float gain, const uint8_t channel);
-float lms7002m_get_rfe_loopback_lna_db(struct lms7002m_context* self, const uint8_t channel);
+lime_Result lms7002m_set_rfe_loopback_lna_db(struct lms7002m_context* self, const double gain, const uint8_t channel);
+double lms7002m_get_rfe_loopback_lna_db(struct lms7002m_context* self, const uint8_t channel);
 
-lime_Result lms7002m_set_rfetia_db(struct lms7002m_context* self, const float value, const uint8_t channel);
-float lms7002m_get_rfetia_db(struct lms7002m_context* self, const uint8_t channel);
+lime_Result lms7002m_set_rfetia_db(struct lms7002m_context* self, const double value, const uint8_t channel);
+double lms7002m_get_rfetia_db(struct lms7002m_context* self, const uint8_t channel);
 
-lime_Result lms7002m_set_trfpad_db(struct lms7002m_context* self, const float value, const uint8_t channel);
-float lms7002m_get_trfpad_db(struct lms7002m_context* self, const uint8_t channel);
+lime_Result lms7002m_set_trfpad_db(struct lms7002m_context* self, const double value, const uint8_t channel);
+double lms7002m_get_trfpad_db(struct lms7002m_context* self, const uint8_t channel);
 
-lime_Result lms7002m_set_trf_loopback_pad_db(struct lms7002m_context* self, const float gain, const uint8_t channel);
-float lms7002m_get_trf_loopback_pad_db(struct lms7002m_context* self, const uint8_t channel);
+lime_Result lms7002m_set_trf_loopback_pad_db(struct lms7002m_context* self, const double gain, const uint8_t channel);
+double lms7002m_get_trf_loopback_pad_db(struct lms7002m_context* self, const uint8_t channel);
 
 lime_Result lms7002m_set_path_rfe(struct lms7002m_context* self, const uint8_t path);
 uint8_t lms7002m_get_path_rfe(struct lms7002m_context* self);
@@ -90,18 +90,18 @@ uint8_t lms7002m_get_band_trf(struct lms7002m_context* self);
 
 lime_Result lms7002m_set_path(struct lms7002m_context* self, bool isTx, uint8_t channel, uint8_t path);
 
-float lms7002m_get_reference_clock_tsp(struct lms7002m_context* self, bool isTx);
+double lms7002m_get_reference_clock_tsp(struct lms7002m_context* self, bool isTx);
 bool lms7002m_get_cgen_locked(struct lms7002m_context* self);
 bool lms7002m_get_sx_locked(struct lms7002m_context* self, bool isTx);
 
 lime_Result lms7002m_tune_vco(struct lms7002m_context* self, enum lms7002m_vco_type module);
 
-float lms7002m_get_frequency_sx(struct lms7002m_context* self, bool isTx);
+double lms7002m_get_frequency_sx(struct lms7002m_context* self, bool isTx);
 
-lime_Result lms7002m_set_nco_frequency(struct lms7002m_context* self, bool isTx, const uint8_t index, float freq_Hz);
-float lms7002m_get_nco_frequency(struct lms7002m_context* self, bool isTx, const uint8_t index);
+lime_Result lms7002m_set_nco_frequency(struct lms7002m_context* self, bool isTx, const uint8_t index, double freq_Hz);
+double lms7002m_get_nco_frequency(struct lms7002m_context* self, bool isTx, const uint8_t index);
 
-lime_Result lms7002m_set_nco_phase_offset_for_mode_0(struct lms7002m_context* self, bool isTx, float angle_deg);
+lime_Result lms7002m_set_nco_phase_offset_for_mode_0(struct lms7002m_context* self, bool isTx, double angle_deg);
 
 #ifdef __cplusplus
 }
